refactor(iv): add type parameter
